+++ docs/CHANGELOG.md
@@ -2,7 +2,6 @@
 
 - Refine selectors <<< HALTED - will continue later >>>
 
-- Need to implement live updates for color picker
 - Important! factor will display on YearView specifically (I haven't yet decided the details on how to implement)
 
 - Change time selector graphic display to show selected values in the center
@@ -11,7 +10,6 @@
 - Settings tab
 - 'Today' button to move back to current day's MonthView instantly
 - Display daily events in MonthView
-- Display month & dates properly in YearView (Button -> BoxLayout w/ ButtonBehavior?)
 - Fix WeekView (looks lame atm)
 - Korean support, fonts
 - Colors, Graphics
@@ -24,6 +22,7 @@
 # \_ADDED:
 
 - COLOUR_RGBA_SELECTED global variable now controls colours displayed for selected values
+- Colour Picker: "Selected Colour" text previews the selected colour in real-time
 
 # \_CHANGED:
 
@@ -31,6 +30,7 @@
 - Time Selector: Added OK / Cancel buttons to make saving / cancelling time selection more straightforward
   - Closing Time selector in any method other than clicking OK will result in just closing the popup without saving
 - Time Selector: Now displays currently selected time correctly when opened
+- Window size adjusted to iPhone 13 (half-scale) resolution
 
 # 2025/04/15